requirements.txt
python-telegram-bot
requests
matplotlib